MotionCanvas_Agent/my-animation/requirement.txt
python-dotenv
google-generativeai
pillow
numpy
opencv-python
pytesseract